+++ CHANGELOG.md
@@ -30,7 +30,8 @@ Only major refactoring improvements or functional changes should be noted.
 #### Changed
 - Catalogs page is now form
 #### New
-- Events datatables/sorting and form/filtering 
+- Events form/filtering by duration and NConf
+- Events datatables/sorting
 - Admin page adds catalogs to db
 
 ## [0.18.0] - 2024-09-27 - Pipeline optimizations